requirements.txt
kubernetes
google-generativeai
python-dotenv
rich
pyfiglet

